Fix a serious bug
I decided to actually run my code, and I found that there
was a bug. 😔

Post mortem:

The function `select` is using `integerToFin` and `restrict`
to compute its result. This is a modest performance
optimization so as to not degrade runtime complexity of
various functions to O(n).

The function definition was wrong (see diff).

This commit includes a fix for the existing code. It also
includes an inductive definition of `select` in comments.

Ideally we would use [idris-quickcheck][1] to test that
the fast implementation matches the behavior of the safer
(inductive) code.

[1]: https://github.com/david-christiansen/idris-quickcheck