Simplify implementations of `Fits`
`OrderedVect`, `LeftistHeap`, and `LazyPairingHeap` previously all
had a variant on a `Fits` ADT. This was a proof obligation to
demonstrate that an element belonged in a particular place in a
data structure.

Turning the various `Fits` ADTs into a function that would either
return a type `()` or a type of form `LTE _ _` seems to work better
in that a lot of 'boring' code can be removed.