Use Rekenaar tactics
Given two equavalent terms in `(Nat, 0, +)`, Rekenaar [1] can prove that
the terms are equal. In this proof-of-concept commit, Rekenaar tactics are
applied everywhere they fit.

There was only one location where applying Rekenaar failed. This code is
marked with a `NOTE` comment. However, it's not clear whether this is
due to a bug in Rekenaar or in Idris.

Note that a case can be made that Rekenaar is best suited for solving
more complex tasks. For example, using Rekenaar to replace a single
`plusCommutative` call might be considered overkill by some users.

[1] https://github.com/jdevuyst/rekenaar